[FE-127] Moved the AccountSelector and BalanceAnnotation componenets to the Header component.
